requirements.txt
# YOLOv8试卷图像分割系统依赖包
# 安装命令: pip install -r requirements.txt

# 核心深度学习框架
ultralytics>=8.0.0          # YOLOv8官方库
# torch>=1.13.0               # PyTorch深度学习框架
# torchvision>=0.14.0         # PyTorch计算机视觉库

# 图像处理
opencv-python>=4.7.0        # OpenCV图像处理库
Pillow>=9.0.0               # Python图像库
numpy>=1.21.0               # 数值计算库

# 数据处理和可视化
matplotlib>=3.5.0           # 绘图库
seaborn>=0.11.0             # 统计可视化库
pandas>=1.4.0               # 数据分析库
scipy>=1.8.0                # 科学计算库

# 配置文件处理
PyYAML>=6.0                 # YAML配置文件解析

# 进度条和日志
tqdm>=4.64.0                # 进度条显示
loguru>=0.6.0               # 高级日志库

# 数据集处理
pycocotools>=2.0.4          # COCO数据集工具

# 图像增强（可选）
albumentations>=1.3.0       # 图像增强库

# 模型优化（可选）
onnx>=1.12.0                # ONNX模型格式
onnxruntime>=1.12.0         # ONNX运行时

# 开发和测试工具（可选）
pytest>=7.0.0               # 测试框架
flake8>=4.0.0               # 代码风格检查
black>=22.0.0               # 代码格式化

# Jupyter支持（可选）
jupyter>=1.0.0              # Jupyter Notebook
ipywidgets>=7.7.0           # Jupyter小部件

# 其他实用工具
requests>=2.28.0            # HTTP请求库
psutil>=5.9.0               # 系统信息库
colorama>=0.4.4             # 终端颜色输出